create config moved to tools
